third_party/WebKit/LayoutTests/fast/selectors/style-sharing-attribute-selector-expected.txt

Style sharing based on attribute selectors

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Attribute selector in nested functional syntax
PASS getComputedStyle(firstDiv).color is black
PASS getComputedStyle(secondDiv).color is green
PASS internals.isSharingStyle(firstDiv, secondDiv) is false
PASS successfullyParsed is true

TEST COMPLETE
This text should be black
This text should be green
